Add back windows support
- the windows wireguard utility requires a config file to bring
  up the wg interface. It is only there to bind the private key
  to the wg0 interface. This also used it to add ip to the iface.
  Peers are all managed with the same library across all platforms
  now.
- broke out interface operations into a seperate file.
