+++ internal/apex/route.go
@@ -1,30 +1,26 @@
 package apex
 
 import (
-	"fmt"
 	"net"
-
-	"go.uber.org/zap"
 )
 
 // handlePeerRoute when a new configuration is deployed, delete/add the peer allowedIPs
 func (ax *Apex) handlePeerRoute(wgPeerConfig wgPeerConfig) {
 	switch ax.os {
 	case Darwin.String():
-		// Darwin maps to a utunX address which needs to be discovered
-		netName, err := getInterfaceByIP(net.ParseIP(ax.wgLocalAddress))
+		// Darwin maps to a utunX address which needs to be discovered (currently hardcoded to utun8)
+		devName, err := getInterfaceByIP(net.ParseIP(ax.wgLocalAddress))
 		if err != nil {
 			ax.logger.Debugf("failed to find the darwin interface with the address [ %s ] %v", ax.wgLocalAddress, err)
 		}
 		// If child prefix split the two prefixes (host /32 and child prefix
 		for _, allowedIP := range wgPeerConfig.AllowedIPs {
-			_, err := RunCommand("route", "-q", "-n", "delete", "-inet", allowedIP, "-interface", netName)
+			_, err := RunCommand("route", "-q", "-n", "delete", "-inet", allowedIP, "-interface", devName)
 			if err != nil {
 				ax.logger.Debugf("no route deleted: %v", err)
 			}
-			_, err = RunCommand("route", "-q", "-n", "add", "-inet", allowedIP, "-interface", netName)
-			if err != nil {
-				ax.logger.Debugf("child prefix route add failed: %v", err)
+			if err := AddRoute(allowedIP, devName); err != nil {
+				ax.logger.Debugf("%v", err)
 			}
 		}
 
@@ -40,55 +36,36 @@ func (ax *Apex) handlePeerRoute(wgPeerConfig wgPeerConfig) {
 				}
 			}
 		}
-	}
-}
 
-// addLinuxChildPrefixRoute check if the prefix exists and add it if not
-func addLinuxChildPrefixRoute(prefix string) error {
-	routeExists, err := RouteExists(prefix)
-	if err != nil {
-		return err
-	}
-
-	if !routeExists {
-		if err := AddRoute(prefix, wgIface); err != nil {
-			return fmt.Errorf("route add failed: %v", err)
+	case Windows.String():
+		for _, allowedIP := range wgPeerConfig.AllowedIPs {
+			if err := AddRoute(allowedIP, wgIface); err != nil {
+				ax.logger.Debugf("route add failed: %v", err)
+			}
 		}
 	}
-	return nil
 }
 
-// addDarwinChildPrefixRoute todo: check if it exists
-func addDarwinChildPrefixRoute(logger *zap.SugaredLogger, prefix string) error {
-	_, err := RunCommand("route", "-q", "-n", "add", "-inet", prefix, "-interface", darwinIface)
-	if err != nil {
-		logger.Debugf("child-prefix route was added: %v", err)
+func (ax *Apex) addChildPrefixRoute(childPrefix string) {
+	var dev string
+
+	if ax.os == Darwin.String() {
+		dev = darwinIface
+	} else {
+		dev = wgIface
 	}
-	return nil
-}
 
-func (ax *Apex) addChildPrefixRoute(ChildPrefix string) {
-	routeExists, err := RouteExists(ChildPrefix)
+	routeExists, err := RouteExists(childPrefix)
 	if err != nil {
 		ax.logger.Warn(err)
 	}
 
-	if ax.os == Linux.String() && routeExists {
-		ax.logger.Debugf("unable to add the child-prefix route [ %s ] as it already exists on this linux host", ChildPrefix)
+	if routeExists {
+		ax.logger.Debugf("unable to add the child-prefix route [ %s ] as it already exists on this linux host", childPrefix)
 		return
 	}
 
-	if ax.os == Linux.String() {
-		if err := addLinuxChildPrefixRoute(ChildPrefix); err != nil {
-			ax.logger.Infof("error adding the child prefix route: %v", err)
-		}
-	}
-
-	// add osx child prefix
-	if ax.os == Darwin.String() {
-		if err := addDarwinChildPrefixRoute(ax.logger, ChildPrefix); err != nil {
-			// TODO: setting to debug until the child prefix is looked up on Darwin
-			ax.logger.Debugf("error adding the child prefix route: %v", err)
-		}
+	if err := AddRoute(childPrefix, dev); err != nil {
+		ax.logger.Infof("error adding the child prefix route: %v", err)
 	}
 }
